--- custom_model_runner/datarobot_drum/cmrunner/exceptions.py
@@ -1,16 +1,0 @@
-class CMRunnerException(Exception):
-    """Base cmrunner exception"""
-
-    pass
-
-
-class CMRunnerCommonException(CMRunnerException):
-    """Raised in case of common errors in cmrunner"""
-
-    pass
-
-
-class CMRunnerPerfTestTimeout(CMRunnerException):
-    """Raised when the perf-test case takes too long"""
-
-    pass
