+++ custom_model_runner/datarobot_drum/drum/main.py
@@ -35,8 +35,8 @@ import os
 import sys
 import signal
 import requests
-from datarobot_drum.cmrunner.args_parser import CMRunnerArgsRegistry
-from datarobot_drum.cmrunner.cmrunner import CMRunner
+from datarobot_drum.drum.args_parser import CMRunnerArgsRegistry
+from datarobot_drum.drum.drum import CMRunner
 
 
 def main():
