+++ custom_model_runner/datarobot_drum/resource/components/Python/python_predictor/python_predictor.py
@@ -2,12 +2,12 @@ import logging
 import sys
 
 from mlpiper.components.connectable_component import ConnectableComponent
-from datarobot_drum.cmrunner.common import (
+from datarobot_drum.drum.common import (
     LOGGER_NAME_PREFIX,
     POSITIVE_CLASS_LABEL_ARG_KEYWORD,
     NEGATIVE_CLASS_LABEL_ARG_KEYWORD,
 )
-from datarobot_drum.cmrunner.model_adapter import PythonModelAdapter
+from datarobot_drum.drum.model_adapter import PythonModelAdapter
 
 logger = logging.getLogger(LOGGER_NAME_PREFIX + "." + __name__)
