+++ custom_model_runner/datarobot_drum/resource/components/Python/r_predictor/r_predictor.py
@@ -3,7 +3,7 @@ import numpy
 import os
 import pandas as pd
 
-from datarobot_drum.cmrunner.common import LOGGER_NAME_PREFIX
+from datarobot_drum.drum.common import LOGGER_NAME_PREFIX
 from mlpiper.components.connectable_component import ConnectableComponent
 
 logger = logging.getLogger(LOGGER_NAME_PREFIX + "." + __name__)
