+++ custom_model_runner/setup.py
@@ -1,8 +1,8 @@
 from setuptools import setup, find_packages
 import os
 
-from datarobot_drum.cmrunner.description import version, project_name
-from datarobot_drum.cmrunner.common import extra_deps, SupportedFrameworks
+from datarobot_drum.drum.description import version, project_name
+from datarobot_drum.drum.common import extra_deps, SupportedFrameworks
 
 # The directory containing this file
 root = os.path.dirname(os.path.abspath(__file__))
