+++ jenkins/test3_mlpiper_custom_models.sh
@@ -15,7 +15,7 @@ git clone git@github.com:datarobot/mlpiper.git
 
 echo "Running tests inside docker:"
 cd $FULL_PATH_CODE_DIR || exit 1
-ls  ./tests/cmrunner/cmrun-tests.sh
+ls  ./tests/drum/cmrun-tests.sh
 
 unameOut="$(uname -s)"
 case "${unameOut}" in
@@ -59,7 +59,7 @@ docker run -i \
       --workdir $FULL_PATH_CODE_DIR \
       -i $TERMINAM_OPTION\
       $DOCKER_IMAGE \
-      ./tests/cmrunner/cmrun-tests.sh
+      ./tests/drum/cmrun-tests.sh
 
 TEST_RESULT=$?
